+++ src/main/java/site/scalarstudios/scalarpower/block/ScalarPowerBlockEntities.java
@@ -3,6 +3,7 @@ package site.scalarstudios.scalarpower.block;
 import site.scalarstudios.scalarpower.ScalarPower;
 import site.scalarstudios.scalarpower.machines.alloysmelter.AlloySmelterBlockEntity;
 import site.scalarstudios.scalarpower.machines.battery.BatteryBlockEntity;
+import site.scalarstudios.scalarpower.machines.battery.CreativeBatteryBlockEntity;
 import site.scalarstudios.scalarpower.machines.battery.SteelBatteryBlockEntity;
 import site.scalarstudios.scalarpower.machines.generator.coal.CoalGeneratorBlockEntity;
 import site.scalarstudios.scalarpower.machines.generator.barometric.BarometricGeneratorBlockEntity;
@@ -91,6 +92,9 @@ public final class ScalarPowerBlockEntities {
 
     public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<SteelBatteryBlockEntity>> STEEL_BATTERY = BLOCK_ENTITY_TYPES
             .register("steel_battery", () -> new BlockEntityType<>(SteelBatteryBlockEntity::new, ScalarPowerBlocks.STEEL_BATTERY.get()));
+
+    public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<CreativeBatteryBlockEntity>> CREATIVE_BATTERY = BLOCK_ENTITY_TYPES
+            .register("creative_battery", () -> new BlockEntityType<>(CreativeBatteryBlockEntity::new, ScalarPowerBlocks.CREATIVE_BATTERY.get()));
 
     public static final DeferredHolder<BlockEntityType<?>, BlockEntityType<CopperWireBlockEntity>> COPPER_WIRE = BLOCK_ENTITY_TYPES
             .register("copper_wire", () -> new BlockEntityType<>(CopperWireBlockEntity::new, ScalarPowerBlocks.COPPER_WIRE.get()));